Block scrobbling internet radio streams
Internet radio streams will report a duration of 0, leading to division
by zero errors. To solve this, I've made it so that mpd_wait_for_play
will block on any tracks that have a duration of 0. However, the old
function would immediatly return true whenever the client state was
'play', so I've had to refactor it to allow for this.

Now, it's written as a loop that will repeat until mpd is playing a
valid (ie. duration not 0) track, that blocks until the client changes
each iteration. There is a boolean value that tracks whether a block was
needed, so that log messages aren't written every time the function is
called, maintaining previous behaviour.